RDKCOM-2660: REFPLTB-1420 : hotspot - Resolve all build warnings
Reason for change: Have to clear all warnings with respect to 64bit architecture
Test Procedure: build should be success on both 64bit and 32bit arch
Risks: Low

(cherry picked from commit 423302777fa9de4783741f36e45692ee3c69a3df)

+++ source/HotspotApi/HotspotApi.c
@@ -706,7 +706,7 @@ int confirmVap(){
            CcspTraceError(("HOTSPOT_LIB : hotspot.json file not available in tmp  %s \n", __FUNCTION__));
            memset((char *)execRetVal,0,sizeof(Err));
            execRetVal->ErrorCode = BLOB_EXEC_FAILURE;
-           return (int)execRetVal;
+           return (intptr_t)execRetVal;
      }
      memset(Buf, '\0', sizeof(Buf));
 //PRASH: Lock /nvram/hotspot.json before copying
